requirements.txt
google-cloud-aiplatform==1.56.0
google-cloud-storage==2.18.2
kfp==2.7.0
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
